hackthebox

+++ nixos/hackthebox.sh
@@ -2,11 +2,19 @@
 
 set -e
 
+user=siddharthist
 # sudo passwd siddharthist pass
 # sudo passwd vagrant pass
-rm -rf dots/ || true
-git clone --branch master --depth 1 https://github.com/langston-barrett/dots
-cd dots/
+
 # TODO: debug this
-HOME=/home/siddharthist bash run.sh || true
-sudo chown -R siddharthist:siddharthist /home/siddharthist
+if [[ ! -d /home/${user}/.config ]]; then
+  rm -rf dots/ || true
+  git clone --branch master --depth 1 https://github.com/langston-barrett/dots
+  cd dots/
+  HOME=/home/${user} bash run.sh || true
+fi
+
+sudo chown -R ${user}:${user} /home/${user}
+# TODO: delete me
+sudo nix-env -i openvpn
+sudo openvpn /etc/nixos/host/*.ovpn & disown
